Renamed DummyEventStore to FakeEventStore to match other mocking-related types
